Update Snake.java Adding logic for bomb in snake...
This logic pinpoints the bomb segment and removes it so that it only removes bombs when a segment within the snake contains the bomb.  There is a bug however in my storing of the snake direction.  Once, that is fixed the bomb should shoot across the screen as the shooting vector logic is correct.  It just needs the direction the snake is facing