Update SnakeGame.java Pinpointed bomb not shooting issue...
Now, I am at a point where everything is coming back correct as far as bomb direction vector calculations with respect to position and direction of the snake.  However, the updat game loop is spawning the bomb offscreen.  So it is  something with spawning/hiding of the bomb.  I am very close to getting it to work