Update Bomb.java Got the bomb flags now properly set but..
The bomb draws in the snakes mouth when I press the triggerbutton.  Just now need the drawing to go across the screen.

Also, bombUpdate is showing invisible bomb position following the vector path through the bitmap, so this is progress.  Once I acheive this, everything else will be a easy